change LionTest
